+++ include/sha1.h
@@ -5,9 +5,9 @@
 
 void do_sha1_first_block(uint8_t data[128], uint32_t state[5]);
 
-void do_sha1_second_block_cpu(uint8_t data[128], size_t len, uint32_t state[5], uint8_t hash[20]);
+void do_sha1_second_block_with_cpu_ext(uint8_t data[128], size_t len, const uint32_t state[5], uint32_t hash[5]);
 
-void do_sha1_second_block_software(uint8_t data[128], size_t len, uint32_t state[5], uint8_t hash[20]);
+void do_sha1_second_block_without_cpu_ext(uint8_t data[128], size_t len, const uint32_t state[5], uint32_t hash[5]);
 
 void sha1_compress_cpu(uint32_t digest[5], const uint8_t *block);