+++ src/ts3idcrunch.c
@@ -70,23 +70,33 @@ static void *stats_start(void *arg) {
     return NULL;
 }
 
-static void *worker_start_software(void *arg) {
-    debug_printf("> worker_start_software(%p)\n", arg);
+static void *worker_start_without_cpu_ext(void *arg) {
+    debug_printf("> worker_start_without_cpu_ext(%p)\n", arg);
     worker_settings *settings = arg;
     uint32_t first_block_state[5]  __attribute__((aligned (16)));
     do_sha1_first_block(settings->pubkey, first_block_state);
     size_t pubkey_len = settings->pubkey_len;
-    uint8_t hash[SHA_DIGEST_LENGTH];
+    uint32_t hash[5];
     uint8_t level_bits_short_circuit =
             settings->level % 8 == 0 ? settings->level : settings->level - (settings->level % 8);
-    debug_printf("  worker_start_software: level_bits_short_circuit=%u\n", level_bits_short_circuit);
+    debug_printf("  worker_start_without_cpu_ext: level_bits_short_circuit=%u\n", level_bits_short_circuit);
     // no logging after this point, performance sensitive!
     while (!do_stop) {
         uint64_t value = atomic_fetch_add(&counter, settings->block_size);
         uint64_t bounds = value + settings->block_size;
+        size_t data_len = append_counter(settings->pubkey, pubkey_len, value);
+        if (data_len > 125) {
+            fprintf(stdout, "You've reached the end the calculating abilities of this tool.\n");
+            fprintf(stdout, "Continuing here makes no sense as the hashrate would halve.\n");
+            fprintf(stdout, "Abort computing process...\n");
+            do_stop = true;
+            if (settings->one_shot) {
+                pthread_mutex_unlock(&keypress_lock);
+                break;
+            }
+        }
         for (uint64_t i = value; i < bounds; i++) {
-            size_t data_len = append_counter(settings->pubkey, pubkey_len, i);
-            do_sha1_second_block_software(settings->pubkey, data_len, first_block_state, hash);
+            do_sha1_second_block_without_cpu_ext(settings->pubkey, data_len, first_block_state, hash);
             uint8_t calc_level = leading_zero_bits(hash, level_bits_short_circuit);
             if (calc_level >= settings->level) {
                 if (results[calc_level] == 0) {
@@ -101,29 +111,40 @@ static void *worker_start_software(void *arg) {
                     break;
                 }
             }
+            data_len = increment_counter(settings->pubkey, pubkey_len, data_len);
         }
     }
-    debug_printf("< worker_start_software(): %p\n", NULL);
+    debug_printf("< worker_start_without_cpu_ext(): %p\n", NULL);
     return NULL;
 }
 
-static void *worker_start_cpu(void *arg) {
-    debug_printf("> worker_start_cpu(%p)\n", arg);
+static void *worker_start_with_cpu_ext(void *arg) {
+    debug_printf("> worker_start_with_cpu_ext(%p)\n", arg);
     worker_settings *settings = arg;
     uint32_t first_block_state[5]  __attribute__((aligned (16)));
     do_sha1_first_block(settings->pubkey, first_block_state);
     size_t pubkey_len = settings->pubkey_len;
-    uint8_t hash[SHA_DIGEST_LENGTH];
+    uint32_t hash[5];
     uint8_t level_bits_short_circuit =
             settings->level % 8 == 0 ? settings->level : settings->level - (settings->level % 8);
-    debug_printf("  worker_start_cpu: level_bits_short_circuit=%u\n", level_bits_short_circuit);
+    debug_printf("  worker_start_with_cpu_ext: level_bits_short_circuit=%u\n", level_bits_short_circuit);
     // no logging after this point, performance sensitive!
     while (!do_stop) {
         uint64_t value = atomic_fetch_add(&counter, settings->block_size);
         uint64_t bounds = value + settings->block_size;
+        size_t data_len = append_counter(settings->pubkey, pubkey_len, value);
+        if (data_len > 125) {
+            fprintf(stdout, "You've reached the end the calculating abilities of this tool.\n");
+            fprintf(stdout, "Continuing here makes no sense as the hashrate would halve.\n");
+            fprintf(stdout, "Abort computing process...\n");
+            do_stop = true;
+            if (settings->one_shot) {
+                pthread_mutex_unlock(&keypress_lock);
+                break;
+            }
+        }
         for (uint64_t i = value; i < bounds; i++) {
-            size_t data_len = append_counter(settings->pubkey, pubkey_len, i);
-            do_sha1_second_block_cpu(settings->pubkey, data_len, first_block_state, hash);
+            do_sha1_second_block_with_cpu_ext(settings->pubkey, data_len, first_block_state, hash);
             uint8_t calc_level = leading_zero_bits(hash, level_bits_short_circuit);
             if (calc_level >= settings->level) {
                 if (results[calc_level] == 0) {
@@ -138,9 +159,10 @@ static void *worker_start_cpu(void *arg) {
                     break;
                 }
             }
+            data_len = increment_counter(settings->pubkey, pubkey_len, data_len);
         }
     }
-    debug_printf("< worker_start_cpu(): %p\n", NULL);
+    debug_printf("< worker_start_with_cpu_ext(): %p\n", NULL);
     return NULL;
 }
 
@@ -238,9 +260,9 @@ static bool start_workers(uint8_t threads, worker_settings settings[threads],
     bool result = true;
     void *(*worker_func)(void *);
     if (check_for_intel_sha_extensions()) {
-        worker_func = worker_start_cpu;
+        worker_func = worker_start_with_cpu_ext;
     } else {
-        worker_func = worker_start_software;
+        worker_func = worker_start_without_cpu_ext;
     }
     for (uint8_t i = 0; i < threads; i++) {
         memset(&settings[i], 0x00, sizeof(struct worker_settings_t));