+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [v1.0.2] - 2020-02-01
+
 * Build: Use cmake module for upx
 * ts3idcrunch: Enhance incrementing the counter (~10% hashrate improvement)
 * ts3idcrunch: Add support for longer counters
@@ -21,7 +23,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * Initial release
 
-[Unreleased]:  https://github.com/bratkartoffel/ts3idtools/compare/v1.0.1...HEAD
+[Unreleased]:  https://github.com/bratkartoffel/ts3idtools/compare/v1.0.2...HEAD
+
+[v1.0.2]:      https://github.com/bratkartoffel/ts3idtools/compare/v1.0.1...v1.0.2
 
 [v1.0.1]:      https://github.com/bratkartoffel/ts3idtools/compare/v1.0.0...v1.0.1